backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
groq
PyPDF2
langchain-groq
python-dotenv
python-dotenv
langchain
langchain-core
